Fix for #26 - indexing of queued items
The problem is that marking anchor as ignored results in not adding it
to the images queue, hence the queue is shorter than total number of
images. The initial processing goes through images and creates a
function that trigger opening an image by its initial index, but the
index is different in the queue due to ignored images.